[worker] moved sender_to_and_from_slice_works/any_as_u8_slice to main
